Refactor code to generate and store keys as JSON file
The commit refactors the "KeyGenerator" class by adding a new method to generate custom cryptographical keys. It also alters the commands to handle JSON files for storing these keys. This improves the quality of the code by making it simpler and easier to use. It also provides a neater way to store and distribute keys for use in cryptographic processes. The Cipher and KeyGenerator classes have also been renamed to KeyGeneratorCommand and KeyGenerator, respectively, to improve clarity on their functions.